feat(nce): expose CLI via NAPI runCli + ship nce bin shim
